#23 move extension scripts to extension folder
