+++ src/extension/ContentScript/shouldInject.js
