+++ src/extension/InpageScript/donation.js
