--- src/styles/_fonts.scss
@@ -1,1 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Nunito:400,600");
