--- src/styles/_reset.scss
@@ -1,10 +1,0 @@
-@import '~advanced-css-reset/dist/reset.css';
-
-// Add your custom reset rules here
-
-* {
-  margin: 0;
-  padding: 0;
-  border: 0;
-  outline: 0;
-}
