--- src/styles/_variables.scss
@@ -1,24 +1,0 @@
-// colors
-$black: #0d0d0d;
-$white: #ffffff;
-$skyBlue: #8892b0;
-$greyWhite: #f3f3f3;
-
-// fonts
-$nunito: "Nunito", sans-serif;
-
-// font weights
-$thin: 100;
-$exlight: 200;
-$light: 300;
-$regular: 400;
-$medium: 500;
-$semibold: 600;
-$bold: 700;
-$exbold: 800;
-$exblack: 900;
-
-// other variables
-.d-none {
-  display: none !important;
-}
